Try to fix image rendering
The interactive plotly graphs are not rendered on GitHub - see https://github.com/plotly/plotly.py/issues/931
A work around is to display these graphs as png as well in the Jupyter notebooks.

I also included the folder notebook_img which should allow the notebooks to render the static images saved there.