Remove the parallel download option
The parallel download option with `concurrent.futures` doesn't seem to be feasible because I think/guess that the servers detect the same IP making parallel requests and then they block the IP (?) Anyways, this is not a very big bottleneck for now. For example I saw that downloading 5221 files took 75 min. We have 41465 URLs sampled for download. That extrapolates linearly to 41465*75/5221 ~ 596 min ~ 10 h